prevision/requirements.txt
appdirs
asttokens
attrs
beautifulsoup4
bs4
certifi
charset-normalizer
comm
contourpy
cssselect
cycler
debugpy
decorator
executing
fake-useragent
fonttools
greenlet
h11
holidays
idna
importlib_metadata
ipykernel
ipynb
ipython
ipython_pygments_lexers
jedi
joblib
jupyter_client
jupyter_core
kiwisolver
lxml
matplotlib
matplotlib-inline
missingno
nest-asyncio
numpy
nvidia-nccl-cu12
outcome
packaging
pandas
parse
parso
pexpect
pillow
platformdirs
playwright
prompt_toolkit
psutil
ptyprocess
pure_eval
pyee
Pygments
pyparsing
pyppeteer
pyquery
PySocks
python-dateutil
python-dotenv
pytz
pyzmq
requests
requests-html
scikit-learn
scipy
seaborn
selenium
six
sniffio
sortedcontainers
soupsieve
stack-data
threadpoolctl
tornado
tqdm
traitlets
trio
trio-websocket
typing_extensions
tzdata
urllib3
w3lib
wcwidth
webdriver-manager
websocket-client
websockets
wsproto
xgboost
zipp
